Discretization new framework relacement
git-tfs-id: [https://infovision.visualstudio.com/DefaultCollection]$/RoughSets;C589

+++ RoughSets/Infovision.Data/DataFieldInfo.cs
@@ -75,7 +75,7 @@ namespace Infovision.Data
         public bool IsUnique { get { return this.isUnique; } set { this.isUnique = value; } }        
         public bool IsOrdered { get { return this.isOrdered; } set { this.isOrdered = value; } }
 
-        public double[] Cuts { get; set; }
+        public long[] Cuts { get; set; }
 
         //TODO NumberOfValues property needs to be changed for numeric data
         public int NumberOfValues
